requirements.txt
# Core dependencies (PyTorch should be installed via conda first)
opencv-python>=4.8.0
Pillow>=10.0.0
timm>=0.9.0
einops>=0.6.0
loguru>=0.7.0
scikit-image>=0.21.0

# Model download
gdown>=4.7.0

# Already installed by conda:
# numpy, scipy, matplotlib
